64-bit changes: casts, and ioctl cmds are u_longs
SCCS-vsn: 8.4

+++ sys/net/bpf.c
@@ -35,7 +35,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *      @(#)bpf.c	8.3 (Berkeley) 08/10/94
+ *      @(#)bpf.c	8.4 (Berkeley) 01/09/95
  *
  * static char rcsid[] =
  * "$Header: bpf.c,v 1.33 91/10/27 21:21:58 mccanne Exp $";
@@ -584,7 +584,7 @@ reset_d(d)
 int
 bpfioctl(dev, cmd, addr, flag)
 	dev_t dev;
-	int cmd;
+	u_long cmd;
 	caddr_t addr;
 	int flag;
 {
